Implemented basic digging
Made detructable tiles and a hitbox that destroys them, said hitbox scales up and down to make sure you can destroy tiles you're touching.

Next step is to make it so you dash through tiles when digging, like Shovel Knight Dig but with the added feature of clinging to the tiles after it if there are any.